@@ -2,6 +2,9 @@
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
+// ReSharper disable MemberCanBePrivate.Global
+// ReSharper disable InconsistentNaming
+// ReSharper disable UnusedMember.Global
 
 namespace Sprache
 {
@@ -284,7 +287,7 @@ namespace Sprache
         /// unqualified counterparts.
         /// </para>
         /// </remarks>
-        /// <seealso cref="XOr"/>
+        /// <seealso cref="XOr{T}"/>
         public static Parser<IEnumerable<T>> XMany<T>(this Parser<T> parser)
         {
             if (parser == null) throw new ArgumentNullException(nameof(parser));
@@ -389,7 +392,7 @@ namespace Sprache
 
                            if (i.Memos.ContainsKey(p))
                            {
-                               var pResult = i.Memos[p] as IResult<T>;
+                               var pResult = (IResult<T>)i.Memos[p];
                                if (pResult.WasSuccessful)
                                    return pResult;
                                throw new ParseException(pResult.ToString());
@@ -593,7 +596,7 @@ namespace Sprache
         /// <returns></returns>
         public static Parser<IEnumerable<T>> Until<T, U>(this Parser<T> parser, Parser<U> until)
         {
-            return parser.Except(until).Many().Then(r => until.Return(r));
+            return parser.Except(until).Many().Then(until.Return);
         }
 
         /// <summary>
@@ -610,7 +613,7 @@ namespace Sprache
 
             return i => parser(i).IfSuccess(s =>
                 predicate(s.Value) ? s : Result.Failure<T>(i,
-                    string.Format("Unexpected {0}.", s.Value),
+                    $"Unexpected {s.Value}.",
                     new string[0]));
         }
 
@@ -686,9 +689,9 @@ namespace Sprache
             if (op == null) throw new ArgumentNullException(nameof(op));
             if (operand == null) throw new ArgumentNullException(nameof(operand));
             if (apply == null) throw new ArgumentNullException(nameof(apply));
-            return or(op.Then(opvalue =>
+            return or(op.Then(opValue =>
                           operand.Then(operandValue =>
-                              ChainOperatorRest(apply(opvalue, firstOperand, operandValue), op, operand, apply, or))),
+                              ChainOperatorRest(apply(opValue, firstOperand, operandValue), op, operand, apply, or))),
                       Return(firstOperand));
         }
 
@@ -742,10 +745,10 @@ namespace Sprache
             if (op == null) throw new ArgumentNullException(nameof(op));
             if (operand == null) throw new ArgumentNullException(nameof(operand));
             if (apply == null) throw new ArgumentNullException(nameof(apply));
-            return or(op.Then(opvalue =>
+            return or(op.Then(opValue =>
                         operand.Then(operandValue =>
                             ChainRightOperatorRest(operandValue, op, operand, apply, or)).Then(r =>
-                                Return(apply(opvalue, lastOperand, r)))),
+                                Return(apply(opValue, lastOperand, r)))),
                       Return(lastOperand));
         }
 
