@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Sprache
+﻿namespace Sprache
 {
     /// <summary>
     /// Represents a customizable comment parser.
