@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Sprache.Tests
+﻿namespace Sprache.Tests
 {
     class PosAwareStr : IPositionAware<PosAwareStr>
     {
